Fixes in:
- MVC arquitecture
- Rotation boidModel
- Boid selection now can unselect boids
