+++ Assets/_Main/Scripts/Boids/BoidsModel.cs
@@ -17,7 +17,6 @@ namespace _Main.Scripts.Boids
         private FlockData m_currentFlockData;
         private BoidsView m_view;
         private Rigidbody m_rigidbody;
-        private RaycastHit[] m_obstacleRayHit;
         public Vector3 WantedDir { get; private set; }
         public bool Is2D { get; private set; }
 
@@ -36,7 +35,7 @@ namespace _Main.Scripts.Boids
             
             var l_targetRotation = Quaternion.LookRotation(p_initDir);
             transform.rotation = l_targetRotation;
-            m_obstacleRayHit = new RaycastHit[5];
+            
             m_view.RefreshFlockView(m_currentFlockData);
             Is2D = p_is2D;
             m_controller.Initialize(this);
@@ -51,8 +50,8 @@ namespace _Main.Scripts.Boids
             
             m_rigidbody.MovePosition(transform.position+l_lerpDir.normalized * (p_speed * Time.deltaTime));
             
-            //transform.position += l_lerpDir.normalized * (p_speed * Time.deltaTime);
-            transform.LookAt(transform.position + l_lerpDir);
+            Quaternion l_targetRotation = Quaternion.LookRotation(l_lerpDir);
+            m_rigidbody.MoveRotation(l_targetRotation);
         }
         
         public void MoveWithAcceleration(Vector2 p_dir, float p_accMult)
@@ -72,27 +71,7 @@ namespace _Main.Scripts.Boids
             transform.LookAt(transform.position + l_lerpDir);
         }
 
-        public float GetCurrSpeedBasedOnDistance(float p_decreasePace)
-        {
-            var l_size = Physics.RaycastNonAlloc(transform.position, transform.forward, m_obstacleRayHit, 
-                m_data.GetStatById(BoidsStatsIds.ViewRange), m_data.ObstacleMask);
-            
-            if (l_size < 1)
-                return m_data.GetStatById(BoidsStatsIds.MovementSpeed);
-
-            var l_closestDistanceToObs = m_data.GetStatById(BoidsStatsIds.ViewRange);
-
-            for (int l_i = 0; l_i < l_size; l_i++)
-            {
-                if (m_obstacleRayHit[l_i].distance < l_closestDistanceToObs)
-                {
-                    l_closestDistanceToObs = m_obstacleRayHit[l_i].distance;
-                }
-            }
-            
-            //return (data.TerminalVelocity)/(p_decreasePace+l_closestDistanceToObs);
-            return m_data.GetStatById(BoidsStatsIds.MovementSpeed) / (1 + (float)Math.Pow(p_decreasePace * l_closestDistanceToObs, 2));
-        }
+        
         
         public void ConstrainTo2D()
         {
