+++ Assets/_Main/Scripts/Managers/MyGame.cs
@@ -3,6 +3,7 @@
     public static class MyGame
     {
         public const string RIGHT_CLICK_ID = "RightClick";
+        public const string LEFT_CLICK_ID = "LeftClick";
         public const string CAMERA_MOVEMENT_ID = "CameraMovement";
         public const string CAMERA_ROTATION_ID = "CameraRotation";
         public const string LEFT_SHIFT_ID = "LeftShift";
